refactor(node): share ref-counted guard across delay and ws
